streamlit_gui/requirements.txt
streamlit>=1.28.0
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.0.0
aiohttp>=3.8.0
asyncio-throttle>=1.0.2
pandas>=2.0.0
plotly>=5.15.0
reportlab>=4.0.0
weasyprint>=60.0
jinja2>=3.1.0
markdown>=3.5.0
fpdf2>=2.7.0

# Testing dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
pytest-asyncio>=0.21.0
requests-mock>=1.11.0
coverage>=7.3.0